move textField and isSubmitting states to Header
